Update Interface
Small changes to interface and use templating to remove duplicate ws
code.
