+++ templates/pico/index.html
@@ -51,7 +51,7 @@
 
             
 
-            <div id="dashboard-sidebar-nav" class="navbar-default sidebar hidden" role="navigation">
+            <div id="dashboard-sidebar-nav" class="navbar-default sidebar" role="navigation">
                 <div class="sidebar-nav navbar-collapse">
                     <ul class="nav" id="side-menu">
                         <li>
@@ -81,15 +81,16 @@
         
         
         <div class="page" id="page-dashboard">
-            <div id="page-wrapper" class="full-width">
+            <div id="page-wrapper">
+                <!--
                 <div class="row">
                     <div class="col-lg-2"></div>
                     <div class="col-lg-8">
                         <div style="font-size:9.0em;text-align:center;color:#5bc0de;border: 6px solid #46b8da;border-radius: 4px;width:1000px;margin:40px auto;">Wallflower.Pico</div>
                     </div>
-                    <div class="col-lg-2"></div>
-                    <!-- /.col-lg-12 -->
+                    <div class='col-lg-2 note'></div>
                 </div>
+                -->
                 <!-- /.row -->
                 <div class="row">
                     <div class="col-lg-12">
@@ -210,10 +211,14 @@
                         <!-- /.panel -->
                     </div>
                     <!-- /.col-lg-8 -->
+                    
+                    <div class='col-lg-2 note'></div>
+                    <!-- /.col-lg-2 -->
+                    
                 </div>
                 <!-- /.row -->
                 
-                <div class="row page-view-network">
+                <div class="row hidden page-view-network">
                     <div class="col-lg-8">
                         <div class="panel panel-default">
                             <div class="panel-heading">
@@ -252,6 +257,10 @@
                         <!-- /.panel -->
                     </div>
                     <!-- /.col-lg-8 -->
+                    
+                    <div class='col-lg-2 note'></div>
+                    <!-- /.col-lg-2 -->
+                    
                 </div>
                 <!-- /.row -->
                 
@@ -295,6 +304,10 @@
                         <!-- /.panel -->
                     </div>
                     <!-- /.col-lg-8 -->
+                    
+                    <div class='col-lg-2 note'></div>
+                    <!-- /.col-lg-2 -->
+                    
                 </div>
                 <!-- /.row -->
                 
@@ -461,8 +474,12 @@
                     </div>
                     <!-- /.col-lg-8 -->
 
+                    <div class='col-lg-2 note'></div>
+                    <!-- /.col-lg-2 -->
+                    
                 </div>
                 <!-- /.row -->
+                
                 
 
                 <div class="row">
@@ -567,7 +584,10 @@
 
                     </div>
                     <!-- /.col-lg-8 -->
-
+                    
+                    <div class='col-lg-2 note'></div>
+                    <!-- /.col-lg-2 -->
+                    
                 </div>
                 <!-- /.row -->
 
@@ -725,6 +745,10 @@
                         <!-- /.panel -->
                     </div>
                     <!-- /.col-lg-8 -->
+                    
+                    <div class='col-lg-2 note'></div>
+                    <!-- /.col-lg-2 -->
+                    
                 </div>
                 <!-- /.row -->
                 
@@ -797,7 +821,7 @@
     <!-- /#wrapper -->
 
     <!-- jQuery -->
-    <script src="/js/jquery.min.js"></script>
+    <script src="/js/jquery-2.2.4.min.js"></script>
     
     <!-- Bootstrap Core JavaScript -->
     <script src="/js/bootstrap.min.js"></script>
@@ -816,8 +840,19 @@
     <!-- DateTimePicker Plugin JavaScript -->
     <script src="/js/jquery.datetimepicker.full.js"></script>
     
+    <script>
+    {% if data.enable_ws %}
+    var enable_ws = true;
+    {% else %}
+    var enable_ws = false;
+    {% endif %}
+    </script>
+    
     <!-- Custom Theme JavaScript -->
     <script src="/js/dashboard.js"></script>
+    
+    <!-- Extensions to Custom Theme JavaScript -->
+    <script src="/js/extend_dashboard.js"></script>
     
 </body>
